Addec clean way of converting a double value to an unsinged integer one. This prevented compilation on an old gcc 4.1.2 system.

+++ source/Optimization/TopGrad.cc
@@ -108,7 +108,7 @@ void TopGrad::SolveProblemCommon(const unsigned int iter)
     if(adaptive_)
     {
       // we only remove a maximum of elems_removed_per_iteration_ of elements in one iteration
-      elems_removed_per_iteration_ = elements_ * pivot_percent_;
+      elems_removed_per_iteration_ = (unsigned int) ( (double) elements_ * pivot_percent_);
     }
 
     // value for comparison
